[mempool] network to lcs interface
This commit eliminates proto from the mempool network interface.
