+++ mempool/src/network.rs
@@ -5,14 +5,34 @@
 
 use crate::counters;
 use channel::{libra_channel, message_queues::QueueStyle};
+use libra_types::transaction::SignedTransaction;
 use libra_types::PeerId;
 use network::{
     error::NetworkError,
     peer_manager::{PeerManagerRequest, PeerManagerRequestSender},
-    proto::MempoolSyncMsg,
-    validator_network::{network_builder::NetworkBuilder, NetworkEvents, NetworkSender},
+    protocols::network::{NetworkEvents, NetworkSender},
+    validator_network::network_builder::NetworkBuilder,
     ProtocolId,
 };
+use serde::{Deserialize, Serialize};
+
+/// Container for exchanging transactions with other Mempools
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub enum MempoolSyncMsg {
+    /// Container for sending transactions to peers
+    BroadcastTransactionsRequest(
+        (
+            u64, // timeline ID that is lower than the first txn in this broadcast's transactions
+            u64, // last timeline ID in this broadcast's transactions
+        ),
+        Vec<SignedTransaction>, // transactions
+    ),
+    /// Response for the BroadcastTransactionsRequest confirming receipt
+    BroadcastTransactionsResponse(
+        u64, // first timeline ID of corresponding request for this response
+        u64, // last timeline ID of corresponding request for this response
+    ),
+}
 
 /// Protocol id for mempool direct-send calls
 pub const MEMPOOL_DIRECT_SEND_PROTOCOL: &[u8] = b"/libra/direct-send/0.1.0/mempool/0.1.0";
