+++ network/build.rs
@@ -5,7 +5,6 @@
 fn main() {
     let proto_files = [
         "src/proto/consensus.proto",
-        "src/proto/mempool.proto",
         "src/proto/state_synchronizer.proto",
     ];
