--- network/src/proto/mempool.proto
@@ -1,10 +1,0 @@
-// Copyright (c) The Libra Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-syntax = "proto3";
-
-package mempool;
-
-/* MempoolSyncMsg represents the messages exchanging between nodes to keep
- * transactions in sync. */
-message MempoolSyncMsg { bytes message = 1; }
